refactor: replace mjs vectordb files by ts files (#36)
