+++ templates/components/loaders/typescript/file/loader.ts
