+++ templates/components/vectordbs/typescript/astra/shared.ts
