+++ templates/components/vectordbs/typescript/milvus/generate.ts
@@ -5,8 +5,8 @@ import {
   VectorStoreIndex,
   storageContextFromDefaults,
 } from "llamaindex";
-import { getDocuments } from "./loader.mjs";
-import { checkRequiredEnvVars, getMilvusClient } from "./shared.mjs";
+import { getDocuments } from "./loader";
+import { checkRequiredEnvVars, getMilvusClient } from "./shared";
 
 dotenv.config();
