+++ templates/components/vectordbs/typescript/mongo/shared.ts
