+++ templates/components/vectordbs/typescript/none/shared.ts
