+++ templates/components/vectordbs/typescript/pg/generate.ts
@@ -1,12 +1,17 @@
 /* eslint-disable turbo/no-undeclared-env-vars */
 import * as dotenv from "dotenv";
 import {
-  PineconeVectorStore,
+  PGVectorStore,
   VectorStoreIndex,
   storageContextFromDefaults,
 } from "llamaindex";
-import { getDocuments } from "./loader.mjs";
-import { checkRequiredEnvVars } from "./shared.mjs";
+import { getDocuments } from "./loader";
+import {
+  PGVECTOR_COLLECTION,
+  PGVECTOR_SCHEMA,
+  PGVECTOR_TABLE,
+  checkRequiredEnvVars,
+} from "./shared";
 
 dotenv.config();
 
@@ -14,20 +19,25 @@ async function loadAndIndex() {
   // load objects from storage and convert them into LlamaIndex Document objects
   const documents = await getDocuments();
 
-  // create vector store
-  const vectorStore = new PineconeVectorStore();
+  // create postgres vector store
+  const vectorStore = new PGVectorStore({
+    connectionString: process.env.PG_CONNECTION_STRING,
+    schemaName: PGVECTOR_SCHEMA,
+    tableName: PGVECTOR_TABLE,
+  });
+  vectorStore.setCollection(PGVECTOR_COLLECTION);
+  vectorStore.clearCollection();
 
-  // create index from all the Documentss and store them in Pinecone
+  // create index from all the Documents
   console.log("Start creating embeddings...");
   const storageContext = await storageContextFromDefaults({ vectorStore });
   await VectorStoreIndex.fromDocuments(documents, { storageContext });
-  console.log(
-    "Successfully created embeddings and save to your Pinecone index.",
-  );
+  console.log(`Successfully created embeddings.`);
 }
 
 (async () => {
   checkRequiredEnvVars();
   await loadAndIndex();
   console.log("Finished generating storage.");
+  process.exit(0);
 })();
