+++ templates/components/vectordbs/typescript/pg/shared.ts
