+++ templates/components/vectordbs/typescript/pinecone/generate.ts
@@ -1,17 +1,12 @@
 /* eslint-disable turbo/no-undeclared-env-vars */
 import * as dotenv from "dotenv";
 import {
-  PGVectorStore,
+  PineconeVectorStore,
   VectorStoreIndex,
   storageContextFromDefaults,
 } from "llamaindex";
-import { getDocuments } from "./loader.mjs";
-import {
-  PGVECTOR_COLLECTION,
-  PGVECTOR_SCHEMA,
-  PGVECTOR_TABLE,
-  checkRequiredEnvVars,
-} from "./shared.mjs";
+import { getDocuments } from "./loader";
+import { checkRequiredEnvVars } from "./shared";
 
 dotenv.config();
 
@@ -19,25 +14,20 @@ async function loadAndIndex() {
   // load objects from storage and convert them into LlamaIndex Document objects
   const documents = await getDocuments();
 
-  // create postgres vector store
-  const vectorStore = new PGVectorStore({
-    connectionString: process.env.PG_CONNECTION_STRING,
-    schemaName: PGVECTOR_SCHEMA,
-    tableName: PGVECTOR_TABLE,
-  });
-  vectorStore.setCollection(PGVECTOR_COLLECTION);
-  vectorStore.clearCollection();
+  // create vector store
+  const vectorStore = new PineconeVectorStore();
 
-  // create index from all the Documents
+  // create index from all the Documentss and store them in Pinecone
   console.log("Start creating embeddings...");
   const storageContext = await storageContextFromDefaults({ vectorStore });
   await VectorStoreIndex.fromDocuments(documents, { storageContext });
-  console.log(`Successfully created embeddings.`);
+  console.log(
+    "Successfully created embeddings and save to your Pinecone index.",
+  );
 }
 
 (async () => {
   checkRequiredEnvVars();
   await loadAndIndex();
   console.log("Finished generating storage.");
-  process.exit(0);
 })();
