+++ templates/components/vectordbs/typescript/pinecone/index.ts
@@ -1,14 +1,13 @@
 /* eslint-disable turbo/no-undeclared-env-vars */
 import {
-  ContextChatEngine,
   LLM,
   PineconeVectorStore,
   VectorStoreIndex,
   serviceContextFromDefaults,
 } from "llamaindex";
-import { CHUNK_OVERLAP, CHUNK_SIZE, checkRequiredEnvVars } from "./shared.mjs";
+import { CHUNK_OVERLAP, CHUNK_SIZE, checkRequiredEnvVars } from "./shared";
 
-async function getDataSource(llm: LLM) {
+export async function getDataSource(llm: LLM) {
   checkRequiredEnvVars();
   const serviceContext = serviceContextFromDefaults({
     llm,
@@ -17,13 +16,4 @@ async function getDataSource(llm: LLM) {
   });
   const store = new PineconeVectorStore();
   return await VectorStoreIndex.fromVectorStore(store, serviceContext);
-}
-
-export async function createChatEngine(llm: LLM) {
-  const index = await getDataSource(llm);
-  const retriever = index.asRetriever({ similarityTopK: 5 });
-  return new ContextChatEngine({
-    chatModel: llm,
-    retriever,
-  });
 }
