+++ templates/components/vectordbs/typescript/pinecone/shared.ts
