+++ templates/types/streaming/express/package.json
@@ -27,6 +27,7 @@
     "typescript": "^5.3.2",
     "prettier": "^3.2.5",
     "prettier-plugin-organize-imports": "^3.2.4",
-    "eslint-config-prettier": "^8.10.0"
+    "eslint-config-prettier": "^8.10.0",
+    "ts-node": "^10.9.2"
   }
 }
