+++ templates/types/streaming/express/tsconfig.json
@@ -9,5 +9,10 @@
     "paths": {
       "@/*": ["./*"]
     }
+  },
+  "ts-node": {
+    "compilerOptions": {
+      "module": "commonjs"
+    }
   }
 }
