+++ templates/types/streaming/nextjs/package.json
@@ -43,6 +43,7 @@
     "cross-env": "^7.0.3",
     "prettier": "^3.2.5",
     "prettier-plugin-organize-imports": "^3.2.4",
-    "eslint-config-prettier": "^8.10.0"
+    "eslint-config-prettier": "^8.10.0",
+    "ts-node": "^10.9.2"
   }
 }
